pyproject.toml
[tool.poetry]
name = "bbbdl"
version = "0.1.0"
description = "A downloader for BigBlueButton meetings"
authors = [
    "g.minoccari <g.minoccari@gmail.com>",
    "Stefano Pigozzi <ste.pigozzi@gmail.com>"
]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
beautifulsoup4 = "^4.9.3"
requests = "^2.24.0"
ffmpeg-python = "^0.2.0"
lxml = "^4.6.1"

[tool.poetry.scripts]
bbbdl = 'bbbdl.__main__:download'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
